Move styles into component
This will move styles into each component.
