+++ components/gallery/Gallery.js
@@ -1,6 +1,5 @@
-import { Text, TouchableOpacity, View } from 'react-native';
+import { Text, TouchableOpacity, View, StyleSheet } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { styles } from './styles/GalleryStyles';
 import { data } from './dummyData';
 import GalleryCard from './GalleryCard';
 import MasonryList from '@react-native-seoul/masonry-list';
@@ -28,3 +27,31 @@ export default function Gallery({ navigation }) {
     </SafeAreaView>
   );
 }
+
+const styles = StyleSheet.create({
+  galleryHeader: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderBottomWidth: 1,
+    borderStyle: 'solid',
+    borderColor: 'grey',
+    paddingVertical: 5,
+    marginBottom: 5,
+  },
+  galleryTitle: {
+    textAlign: 'center',
+    fontSize: 28,
+    fontWeight: 'bold',
+  },
+  uploadBtn: {
+    position: 'absolute',
+    right: 0,
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 4,
+  },
+  galleryContent: {
+    paddingHorizontal: 12,
+    alignSelf: 'stretch',
+  },
+});
