--- components/gallery/styles/GalleryCardStyles.js
@@ -1,9 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  itemImage: {
-    resizeMode: 'stretch',
-    borderRadius: 5,
-    margin: 5,
-  },
-});
