--- components/gallery/styles/GalleryStyles.js
@@ -1,29 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  galleryHeader: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderBottomWidth: 1,
-    borderStyle: 'solid',
-    borderColor: 'grey',
-    paddingVertical: 5,
-    marginBottom: 5,
-  },
-  galleryTitle: {
-    textAlign: 'center',
-    fontSize: 28,
-    fontWeight: 'bold',
-  },
-  uploadBtn: {
-    position: 'absolute',
-    right: 0,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 4,
-  },
-  galleryContent: {
-    paddingHorizontal: 12,
-    alignSelf: 'stretch',
-  },
-});
